Removed SSR since Amplify can not start a node file without using Amplify CLI
