+++ .eslintrc.cjs
@@ -2,9 +2,6 @@
 require('@rushstack/eslint-patch/modern-module-resolution')
 
 module.exports = {
-  "env": {
-    "node": true
-  },
   root: true,
   'extends': [
     'plugin:vue/vue3-essential',
