--- pages/index.page.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import HomeView from "@/views/HomeView.vue";
-</script>
-
-<template>
-  <HomeView />
-</template>
-
-<style scoped>
-
-</style>
