--- renderer/_default.page.client.js
@@ -1,18 +1,0 @@
-// Environment: browser
-import { createApp } from './app'
-// Import the app styles. This will also enable Tailwind CSS styles
-import '../src/assets/main.css'
-
-// This render() hook only supports SSR, see https://vite-plugin-ssr.com/render-modes for how to modify render() to support SPA
-async function render(pageContext) {
-  const { Page, pageProps } = pageContext
-  if (!Page) throw new Error('Client-side render() hook expects pageContext.Page to be defined')
-  const app = createApp(Page, pageProps, pageContext)
-  app.mount('#app')
-}
-
-/* To enable Client-side Routing:
-export const clientRouting = true
-// !! WARNING !! Before doing so, read https://vite-plugin-ssr.com/clientRouting */
-
-export { render }
