--- renderer/_default.page.server.js
@@ -1,41 +1,0 @@
-// Environment: server
-import { renderToString } from '@vue/server-renderer'
-import { escapeInject, dangerouslySkipEscape } from 'vite-plugin-ssr/server'
-import {createApp} from './app.js'
-
-async function render(pageContext) {
-  const { Page, pageProps } = pageContext
-  // This render() hook only supports SSR, see https://vite-plugin-ssr.com/render-modes for how to modify render() to support SPA
-  if (!Page) throw new Error('My render() hook expects pageContext.Page to be defined')
-  const app = createApp(Page, pageProps, pageContext)
-
-  const appHtml = await renderToString(app)
-
-  const title = 'Opinion Rate It'
-
-  return escapeInject`<!DOCTYPE html>
-    <html lang="en">
-      <head>
-        <meta charset="UTF-8">
-        <link rel="icon" href="/favicon.ico">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link rel="stylesheet" href="//fonts.googleapis.com/css?family=Roboto:400,500,700,400italic|Material+Icons">
-        <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap" rel="stylesheet">
-        <title>${title}</title>
-      </head>
-      <body>
-        <div id="app">${dangerouslySkipEscape(appHtml)}</div>
-      </body>
-    </html>`
-}
-
-// Example of `pageContext` often passed to the browser
-const passToClient = [
-  'pageProps',
-  'routeParams',
-  // (Deep selection isn't implemented yet; open a GitHub ticket if you want this.)
-  'user.id',
-  'user.name'
-]
-
-export { render, passToClient }
