+++ src/controllers/carouselController.ts
@@ -13,17 +13,13 @@ const useCarousel = (slides: Slide[]) => {
 
   let animateId: number
   const startAnimation = () => {
-    if (!import.meta.env.SSR) {
-      animateId = window.setInterval(() => {
-        next()
-      }, 3000)
-    }
+    animateId = window.setInterval(() => {
+      next()
+    }, 3000)
   }
 
   const stopAnimation = () => {
-    if (!import.meta.env.SSR && animateId) {
-      window.clearInterval(animateId)
-    }
+    window.clearInterval(animateId)
   }
 
   const slideAlias = ref(slides)
