+++ src/controllers/observerController.ts
@@ -1,31 +1,28 @@
 import {nextTick, onBeforeUnmount, onMounted} from "vue";
 
 const useRevealObserver = () => {
-
-    if (!import.meta.env.SSR) {
-        const callback = (entries: any[]) => {
-            entries.forEach(entry => {
-                if (entry.isIntersecting) {
-                    entry.target.classList.add('active')
-                } else {
-                    entry.target.classList.remove('active')
-                }
-            })
-        }
-
-        const observer: IntersectionObserver = new IntersectionObserver(callback)
-
-        onMounted(async () => {
-            await nextTick()
-            document.querySelectorAll('.reveal').forEach(el => {
-                observer.observe(el)
-            })
-        })
-
-        onBeforeUnmount(() => {
-            observer.disconnect()
+    const callback = (entries: any[]) => {
+        entries.forEach(entry => {
+            if (entry.isIntersecting) {
+                entry.target.classList.add('active')
+            } else {
+                entry.target.classList.remove('active')
+            }
         })
     }
+
+    const observer: IntersectionObserver = new IntersectionObserver(callback)
+
+    onMounted(async () => {
+        await nextTick()
+        document.querySelectorAll('.reveal').forEach(el => {
+            observer.observe(el)
+        })
+    })
+
+    onBeforeUnmount(() => {
+        observer.disconnect()
+    })
 }
 
 export {
