+++ tailwind.config.js
@@ -1,9 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    "./renderer/**/*.{vue,js,ts,jsx,tsx}",
+    "./index.html",
     "./src/**/*.{vue,js,ts,jsx,tsx}",
-    "./pages/**/*.{vue,js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {},
