+++ tsconfig.node.json
@@ -5,12 +5,5 @@
     "composite": true,
     "module": "ESNext",
     "types": ["node"]
-  },
-  "ts-node": {
-    "transpileOnly": true,
-    "esm": true,
-    "compilerOptions": {
-      "moduleResolution": "Node16"
-    }
   }
 }
